+++ Basil/include/Platform/OpenGL/OpenGLShader.h
@@ -3,6 +3,8 @@
  */
 #pragma once
 
+#include <unordered_map>
+
 #include "Renderer/Shader.h"
 
 #include <glm/glm.hpp>